+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 # Unreleased
+- add DuckDB::ScalarFunction#set_return_type.
 
 # 1.4.3.0 - 2026-01-10
 - bump duckdb to 1.4.3 on CI.